+++ services/madoc-ts/src/utility/errors/not-found.ts
@@ -1,1 +1,3 @@
-export class NotFound extends Error {}
+export class NotFound extends Error {
+  status = 404;
+}